Add useOmniAutoBSDataHandler
