+++ features/omni-kit/automation/components/index.ts
@@ -5,4 +5,4 @@ export * from './trailing-stop-loss/OmniTrailingStopLossOverviewDetailsSection'
 export * from './trailing-stop-loss/OmniTrailingStopLossOverviewDetailsSectionFooter'
 export * from './partial-take-profit/OmniPartialTakeProfitOverviewDetailsSection'
 export * from './partial-take-profit/OmniPartialTakeProfitOverviewDetailsSectionFooter'
-export * from './OmniAutoBSOverviewDetailsSection'
+export * from './auto-buy-sell/OmniAutoBSOverviewDetailsSection'
