+++ features/omni-kit/automation/components/partial-take-profit/OmniPartialTakeProfitOverviewDetailsSection.tsx
@@ -2,7 +2,7 @@ import { DetailsSection } from 'components/DetailsSection'
 import { DetailsSectionContentCardWrapper } from 'components/DetailsSectionContentCard'
 import { OmniPartialTakeProfitOverviewDetailsSectionFooter } from 'features/omni-kit/automation/components'
 import { resolveActiveOrder } from 'features/omni-kit/automation/helpers'
-import { useOmniPartialTakeProfitDataHandler } from 'features/omni-kit/automation/helpers/useOmniPartialTakeProfitDataHandler'
+import { useOmniPartialTakeProfitDataHandler } from 'features/omni-kit/automation/hooks'
 import { OmniContentCard } from 'features/omni-kit/components/details-section'
 import { useOmniGeneralContext } from 'features/omni-kit/contexts'
 import { useTranslation } from 'next-i18next'
