+++ features/omni-kit/automation/components/trailing-stop-loss/OmniTrailingStopLossOverviewDetailsSection.tsx
@@ -2,7 +2,7 @@ import { DetailsSection } from 'components/DetailsSection'
 import { DetailsSectionContentCardWrapper } from 'components/DetailsSectionContentCard'
 import { OmniTrailingStopLossOverviewDetailsSectionFooter } from 'features/omni-kit/automation/components'
 import { resolveActiveOrder } from 'features/omni-kit/automation/helpers'
-import { useOmniTrailingStopLossDataHandler } from 'features/omni-kit/automation/helpers/useOmniTrailingStopLossDataHandler'
+import { useOmniTrailingStopLossDataHandler } from 'features/omni-kit/automation/hooks'
 import { OmniContentCard } from 'features/omni-kit/components/details-section'
 import { useOmniGeneralContext, useOmniProductContext } from 'features/omni-kit/contexts'
 import { useTranslation } from 'next-i18next'
