+++ features/omni-kit/automation/helpers/index.ts
@@ -3,6 +3,5 @@ export * from './hasActiveProtection'
 export * from './resolveActiveOrder'
 export * from './getRealizedProfit'
 export * from './automationFormatters'
-export * from './useOmniStopLossDataHandler'
 export * from './isOmniAutomationFormEmpty'
 export * from './getOmniAutomationParameters'
