+++ features/omni-kit/automation/hooks/useOmniStopLossDataHandler.tsx
